pyproject.toml
[tool.poetry]
name = "woertle"
version = "0.1.0"
description = ""
authors = ["jkoessle <mail@jonathan-koessler.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
pre-commit = "^4.0.1"
ruff = "^0.8.4"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
